test(init,inspect,team): the scaffolds and reports become spec documents
Twenty-four tests across five directories, twenty-two of them documents.
`init` is the clearest case the format has: one command, two banners, and
a tree of files it either wrote or must not have written. The starter
scaffold's twenty-odd `expect(result.file(…))` lines are now one `files:`
block reading as an inventory — `contains` for the four local-ref schemes,
`absent` for `spwn/hooks.yaml`, `spwn/worlds/default.yaml` and the
root-level `knowledge/` the 2026-04 relocation retired.

`inspect --offline` is deterministic by design, which is why it already
had a golden; three of its specs were probing that same golden by
substring or by counting header lines, and a byte-exact `stdout:` says
all of it at once.

What stays in code: the three scaffold rules that are ABSENCES a regex
must express — no colon-form local ref, no root-level knowledge path, no
pinned runtime backend — bridged onto the scaffolding document so the
session is stated once. And the two docker specs in `skill/`, which spin
a world and read the container back.
