+++ tests/specs/cli/init/bare-gallery-slug.spec.yaml
@@ -1,0 +1,20 @@
+description: a bare catalog slug resolves to the gallery entry
+fixture: $FIXTURES/empty/
+runs:
+    - command: init matrix
+      exit: 0
+      stdout: |4
+            spwn.yaml          # created
+            Worlds added:      matrix
+            Agents added:      neo
+
+            Next:
+              spwn up                        # spawn the world
+      stderr: |4+
+          
+            ✓ Installed example spwn:matrix
+
+
+      files:
+          spwn.yaml: exists
+          spwn/agents/neo/SOUL.md: exists
