+++ tests/specs/cli/init/existing-manifest.spec.yaml
@@ -1,0 +1,10 @@
+description: init errors when spwn.yaml already exists
+fixture: $FIXTURES/single-agent/
+runs:
+    - command: init
+      exit: 1
+      stderr: |4+
+          
+            Error: spwn.yaml already exists at {{workdir}}/spwn.yaml.
+          Use --force to overwrite
+
