--- tests/specs/cli/init/expected/init-already-exists.txt
@@ -1,4 +1,0 @@
-
-  Error: spwn.yaml already exists at {{workdir}}/spwn.yaml.
-Use --force to overwrite
-
