--- tests/specs/cli/init/expected/init-banner.txt
@@ -1,4 +1,0 @@
-
-  ✓ Initialised spwn project demo-project
-
-
