--- tests/specs/cli/init/expected/init-matrix-banner.txt
@@ -1,4 +1,0 @@
-
-  ✓ Installed example spwn:matrix
-
-
