--- tests/specs/cli/init/expected/init-output.txt
@@ -1,15 +1,0 @@
-  Committed:
-    spwn.yaml                      # project manifest (worlds live inline)
-    spwn.lock                      # committed dependency pins (like package-lock.json)
-    spwn/agents/neo/               # starter agent
-    spwn/skills/focus.md           # example local skill  (skill/focus)
-    spwn/tools/greet/              # example local tool   (tool/greet)
-    spwn/hooks/session-banner.yaml # example local hook   (hook/session-banner)
-    spwn/commands/refactor.md      # example slash command (command/refactor)
-
-  Gitignored:
-    .spwn/                         # local state
-
-  Next:
-    spwn auth                      # verify your credentials
-    spwn agent neo                 # open an interactive session with neo
