+++ tests/specs/cli/init/forced-overwrite.spec.yaml
@@ -1,0 +1,28 @@
+description: --force overwrites an existing spwn.yaml
+fixture: $FIXTURES/single-agent/
+runs:
+    - command: init --force --name forced-project
+      exit: 0
+      stdout: |4
+            Committed:
+              spwn.yaml                      # project manifest (worlds live inline)
+              spwn.lock                      # committed dependency pins (like package-lock.json)
+              spwn/agents/neo/               # starter agent
+              spwn/skills/focus.md           # example local skill  (skill/focus)
+              spwn/tools/greet/              # example local tool   (tool/greet)
+              spwn/hooks/session-banner.yaml # example local hook   (hook/session-banner)
+              spwn/commands/refactor.md      # example slash command (command/refactor)
+
+            Gitignored:
+              .spwn/                         # local state
+
+            Next:
+              spwn auth                      # verify your credentials
+              spwn agent neo                 # open an interactive session with neo
+      stderr: |4+
+          
+            ✓ Initialised spwn project forced-project
+
+
+      files:
+          spwn.yaml: { contains: "name: forced-project" }
