+++ tests/specs/cli/init/gallery-example.spec.yaml
@@ -1,0 +1,21 @@
+description: init spwn:matrix installs the bundled example
+fixture: $FIXTURES/empty/
+runs:
+    - command: init spwn:matrix
+      exit: 0
+      stdout: |4
+            spwn.yaml          # created
+            Worlds added:      matrix
+            Agents added:      neo
+
+            Next:
+              spwn up                        # spawn the world
+      stderr: |4+
+          
+            ✓ Installed example spwn:matrix
+
+
+      files:
+          spwn.yaml: exists
+          spwn/agents/neo/agent.yaml: exists
+          spwn/agents/neo/SOUL.md: exists
