+++ tests/specs/cli/init/starter-scaffold.spec.yaml
@@ -1,0 +1,54 @@
+description: init scaffolds the starter project in an empty directory
+fixture: $FIXTURES/empty/
+runs:
+    - command: init --name demo-project
+      exit: 0
+      stdout: |4
+            Committed:
+              spwn.yaml                      # project manifest (worlds live inline)
+              spwn.lock                      # committed dependency pins (like package-lock.json)
+              spwn/agents/neo/               # starter agent
+              spwn/skills/focus.md           # example local skill  (skill/focus)
+              spwn/tools/greet/              # example local tool   (tool/greet)
+              spwn/hooks/session-banner.yaml # example local hook   (hook/session-banner)
+              spwn/commands/refactor.md      # example slash command (command/refactor)
+
+            Gitignored:
+              .spwn/                         # local state
+
+            Next:
+              spwn auth                      # verify your credentials
+              spwn agent neo                 # open an interactive session with neo
+      stderr: |4+
+          
+            ✓ Initialised spwn project demo-project
+
+
+      files:
+          spwn.yaml:
+              contains:
+                  - "name: demo-project"
+                  - "knowledge: ./spwn/knowledge"
+          spwn/agents/neo/agent.yaml:
+              contains:
+                  - skill/focus
+                  - tool/greet
+                  - hook/session-banner
+                  - command/refactor
+          spwn/agents/neo/SOUL.md: exists
+          spwn/agents/neo/AGENTS.md: exists
+          .gitignore: { contains: .spwn }
+          spwn/skills/focus.md: { contains: "name: focus" }
+          spwn/tools/greet/tool.yaml: { contains: "name: greet" }
+          spwn/hooks/session-banner.yaml:
+              contains:
+                  - "event: SessionStart"
+                  - "command:"
+          spwn/commands/refactor.md:
+              contains:
+                  - Refactor
+                  - "description:"
+          spwn/knowledge/.gitkeep: exists
+          spwn/hooks.yaml: absent
+          spwn/worlds/default.yaml: absent
+          knowledge: absent
