+++ tests/specs/cli/init/tool-shaped-entry.spec.yaml
@@ -1,0 +1,14 @@
+description: init refuses a tool-shaped catalog entry that is not gallery-eligible
+fixture: $FIXTURES/empty/
+runs:
+    - command: init qmd
+      exit: 1
+      stderr: |4+
+          
+            Error: "qmd" is not in the catalog.
+          known: macrohard, matrix, paperclip-factory, research-lab, severance, startup
+
+          hint: pass skill/qmd, tool/qmd, hook/qmd, or command/qmd for a local block
+
+      files:
+          spwn.yaml: absent
