+++ tests/specs/cli/init/unknown-backend.spec.yaml
@@ -1,0 +1,11 @@
+description: --backend rejects an unknown runtime with a supported-list hint
+fixture: $FIXTURES/empty/
+runs:
+    - command: init --name bad-demo --backend does-not-exist
+      exit: 1
+      stderr: |4+
+          
+            Error: unknown --backend "does-not-exist" — supported: claude-code, spwn:claude-code, codex, spwn:codex
+
+      files:
+          spwn.yaml: absent
