+++ tests/specs/cli/init/unknown-gallery-slug.spec.yaml
@@ -1,0 +1,14 @@
+description: an unknown slug errors with the catalog hint
+fixture: $FIXTURES/empty/
+runs:
+    - command: init nonesuch
+      exit: 1
+      stderr: |4+
+          
+            Error: "nonesuch" is not in the catalog.
+          known: macrohard, matrix, paperclip-factory, research-lab, severance, startup
+
+          hint: pass skill/nonesuch, tool/nonesuch, hook/nonesuch, or command/nonesuch for a local block
+
+      files:
+          spwn.yaml: absent
