--- tests/specs/cli/inspect/expected/single-agent.txt
@@ -1,15 +1,0 @@
-Name         neo
-Role         worker
-Runtime      claude-code
-World        neo   (○ stopped)
-
-Dependencies (3 direct, 0 transitive)
-  spwn:unix@24.04
-  spwn:git@latest
-  spwn:python@3
-
-Skills (0)
-  -
-
-Hooks (0)
-  -
