+++ tests/specs/cli/inspect/local-tool-ref.spec.yaml
@@ -1,0 +1,20 @@
+description: a local tool dependency renders as a bare name
+fixture: $FIXTURES/mixed-tool-refs/
+runs:
+    - command: inspect --offline
+      exit: 0
+      stdout: |
+          Name         neo
+          Role         worker
+          Runtime      claude-code
+          World        default   (○ stopped)
+
+          Dependencies (2 direct, 0 transitive)
+            spwn:unix@24.04
+            tool/my-local-tool
+
+          Skills (0)
+            -
+
+          Hooks (0)
+            -
