+++ tests/specs/cli/inspect/missing-agent.spec.yaml
@@ -1,0 +1,9 @@
+description: inspect errors cleanly when the named agent is missing
+fixture: $FIXTURES/single-agent/
+runs:
+    - command: inspect ghost --offline
+      exit: 1
+      stderr: |4+
+          
+            Error: agent "ghost" not found in this project
+
