+++ tests/specs/cli/inspect/outside-a-project.spec.yaml
@@ -1,0 +1,9 @@
+description: inspect outside a spwn project names the missing manifest
+fixture: $FIXTURES/empty/
+runs:
+    - command: inspect --offline
+      exit: 1
+      stderr: |4+
+          
+            Error: no spwn.yaml found in {{workdir}} or any parent — run `spwn init` first
+
