+++ tests/specs/cli/inspect/single-agent-report.spec.yaml
@@ -1,0 +1,21 @@
+description: inspect renders a single-agent project end to end
+fixture: $FIXTURES/single-agent/
+runs:
+    - command: inspect --offline
+      exit: 0
+      stdout: |
+          Name         neo
+          Role         worker
+          Runtime      claude-code
+          World        neo   (○ stopped)
+
+          Dependencies (3 direct, 0 transitive)
+            spwn:unix@24.04
+            spwn:git@latest
+            spwn:python@3
+
+          Skills (0)
+            -
+
+          Hooks (0)
+            -
