+++ tests/specs/cli/inspect/usage-page.spec.yaml
@@ -1,0 +1,27 @@
+description: inspect --help prints the command usage
+fixture: $FIXTURES/empty/
+runs:
+    - command: inspect --help
+      exit: 0
+      stdout: |+
+          
+          Inspect a spwn project: one block per agent, showing the
+          resolved dependency tree (with transitive (*)-dedup and composition
+          badges), the skills contributed by tool deps, and the hooks bound
+          to the agent.
+
+          Mirrors the kubectl describe / cargo tree convention: key-value
+          header, section titles with counts, whitespace-separated sections.
+
+          Examples:
+            spwn inspect            # every agent
+            spwn inspect neo        # one agent
+            spwn inspect --offline  # skip live world-status lookup
+
+          Usage:
+            spwn inspect [agent] [flags]
+
+          Flags:
+            -h, --help      help for inspect
+                --offline   Skip live world-status lookup (faster, no Docker calls)
+
