+++ tests/specs/cli/skill/project-local-skill.spec.yaml
@@ -1,0 +1,33 @@
+description: skill new inside a project writes into the project tree
+fixture: $FIXTURES/empty/
+runs:
+    - command: init
+      exit: 0
+      stdout: |4
+            Committed:
+              spwn.yaml                      # project manifest (worlds live inline)
+              spwn.lock                      # committed dependency pins (like package-lock.json)
+              spwn/agents/neo/               # starter agent
+              spwn/skills/focus.md           # example local skill  (skill/focus)
+              spwn/tools/greet/              # example local tool   (tool/greet)
+              spwn/hooks/session-banner.yaml # example local hook   (hook/session-banner)
+              spwn/commands/refactor.md      # example slash command (command/refactor)
+
+            Gitignored:
+              .spwn/                         # local state
+
+            Next:
+              spwn auth                      # verify your credentials
+              spwn agent neo                 # open an interactive session with neo
+      stderr: |4+
+          
+            ✓ Initialised spwn project {{string}}
+
+
+    - command: skill new my-skill
+      exit: 0
+      stderr: |
+          Authored {{workdir}}/spwn/skills/my-skill.md
+          Edit with 'spwn skill edit my-skill' or open the file directly.
+      files:
+          spwn/skills/my-skill.md: exists
