--- tests/specs/cli/team/expected/team-required.txt
@@ -1,3 +1,0 @@
-
-  Error: team name is required
-
